requirements.txt
click==7.1.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Pillow==7.1.1
pkg-resources==0.0.0
Werkzeug==1.0.1
